+++ my-react-app/pages/PageLearn.css
@@ -1,0 +1,11 @@
+iframe{
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    overflow: hidden;
+    width: 600px;
+    height: 450px;
+    border-radius: 5px;
+}
+
