+++ my-react-app/src/components/Panneau/Panneau.tsx
@@ -15,6 +15,7 @@ export default function Panneau () {
         <FontAwesomeIcon className="vis3" icon={faCircleXmark} style={{color: "#ababab",}} />
         <FontAwesomeIcon className="vis4" icon={faCircleXmark} style={{color: "#ababab",}} />
       </article>
+      
     </div>
   );
 };
