+++ my-react-app/src/components/ThreeButtonsPage/ThreeButtonsPage.css
@@ -1,0 +1,30 @@
+.button-activity {
+    width: 70%;
+    height: 4rem;
+    padding: 0.5rem 2rem 0.5rem 2rem;
+    border-radius: 4rem;
+    border: none;
+    font-family: "Monomaniac One";
+    background-color: #E3E3E3;
+    color: #969191;
+    font-size: 1.2rem;
+    cursor: pointer;
+}
+
+.container-buttons {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    gap: 2rem;
+    height: 100vh;
+}
+
+@media screen and (width > 768px){
+    .button-activity {
+        width: 20rem;
+        height: 5rem;
+        font-size: 1.5rem;
+    }
+}
+
